Refactor API settings page and show keys based on abac perms
