+++ studio/components/ui/ProjectSettings/DisplayApiSettings.tsx
@@ -1,11 +1,12 @@
-import { FC } from 'react'
 import { useRouter } from 'next/router'
 import { Input } from '@supabase/ui'
 import { JwtSecretUpdateStatus } from '@supabase/shared-types/out/events'
-import { useJwtSecretUpdateStatus, useProjectSettings } from 'hooks'
-import { SettingsLoadingState } from './SettingsLoadingState'
+import { PermissionAction } from '@supabase/shared-types/out/constants'
+
+import { checkPermissions, useJwtSecretUpdateStatus, useProjectSettings } from 'hooks'
 import { DEFAULT_PROJECT_API_SERVICE_ID } from 'lib/constants'
 import Panel from 'components/ui/Panel'
+import { SettingsLoadingState } from './SettingsLoadingState'
 
 const DisplayApiSettings = () => {
   const router = useRouter()
@@ -21,6 +22,8 @@ const DisplayApiSettings = () => {
     jwtSecretUpdateStatus,
   }: any = useJwtSecretUpdateStatus(ref)
 
+  const canReadAPIKeys = checkPermissions(PermissionAction.READ, 'service_api_keys')
+
   const isNotUpdatingJwtSecret =
     jwtSecretUpdateStatus === undefined || jwtSecretUpdateStatus === JwtSecretUpdateStatus.Updated
   // Get the API service
@@ -30,7 +33,18 @@ const DisplayApiSettings = () => {
   const isApikeysEmpty = apiKeys.length === 0
 
   return (
-    <ApiContentWrapper>
+    <Panel
+      title={
+        <div className="space-y-3">
+          <h5 className="text-base">Project API keys</h5>
+          <p className="text-sm text-scale-1000">
+            Your API is secured behind an API gateway which requires an API Key for every request.
+            <br />
+            You can use the keys below to use Supabase client libraries.
+          </p>
+        </div>
+      }
+    >
       {isProjectSettingsLoading || isJwtSecretUpdateStatusLoading || isApikeysEmpty ? (
         <SettingsLoadingState
           isError={isProjectSettingsError || isJwtSecretUpdateStatusError}
@@ -46,7 +60,10 @@ const DisplayApiSettings = () => {
             }
           >
             <Input
+              readOnly
+              disabled
               layout="horizontal"
+              className="input-mono"
               // @ts-ignore
               label={
                 <>
@@ -63,13 +80,12 @@ const DisplayApiSettings = () => {
                   {x.tags === 'anon' && <code className="text-xs text-code">{'public'}</code>}
                 </>
               }
-              readOnly
-              copy={isNotUpdatingJwtSecret}
-              className="input-mono"
-              disabled
-              reveal={x.tags !== 'anon' && isNotUpdatingJwtSecret}
+              copy={canReadAPIKeys && isNotUpdatingJwtSecret}
+              reveal={x.tags !== 'anon' && canReadAPIKeys && isNotUpdatingJwtSecret}
               value={
-                jwtSecretUpdateStatus === JwtSecretUpdateStatus.Failed
+                !canReadAPIKeys
+                  ? 'You need additional permissions to read API keys'
+                  : jwtSecretUpdateStatus === JwtSecretUpdateStatus.Failed
                   ? 'JWT secret update failed, new API key may have issues'
                   : jwtSecretUpdateStatus === JwtSecretUpdateStatus.Updating
                   ? 'Updating JWT secret...'
@@ -85,26 +101,7 @@ const DisplayApiSettings = () => {
           </Panel.Content>
         ))
       )}
-    </ApiContentWrapper>
-  )
-}
-export default DisplayApiSettings
-
-const ApiContentWrapper: FC = ({ children }) => {
-  return (
-    <Panel
-      title={
-        <div className="space-y-3">
-          <h5 className="text-base">Project API keys</h5>
-          <p className="text-sm text-scale-1000">
-            Your API is secured behind an API gateway which requires an API Key for every request.
-            <br />
-            You can use the keys below to use Supabase client libraries.
-          </p>
-        </div>
-      }
-    >
-      {children}
     </Panel>
   )
 }
+export default DisplayApiSettings
